// FrontEnd/Front_TFC/src/app/interfaces/evento-response-dto.ts
export interface EventoResponseDto {
    idEvento: number
    nombre: string
    descripcion: string
    fechaInicio: string
    horaInicio: string;         
    horaFin: string;  
    estado: string
    destacado: string
    aforoMaximo: number
    minimoAsistencia: number
    precio: number
    imagen: string
    tipo: string
    plazasLibres: number
    zona: string
}